requirements.txt
Flask
psycopg2==2.9.1
python-decouple==3.5
flask_sqlalchemy